Move components to shared
