+++ app/frontend-webapp/components/core/sidebar/SidebarItem.tsx
@@ -2,7 +2,7 @@
 
 import { usePathname } from "next/navigation";
 
-import { Icons } from "@/components/Icons";
+import { Icons } from "@/components/shared/Icons";
 import {
   Accordion,
   AccordionContent,
